scripts/requirements.txt
beautifulsoup4
requests
humanize
python-dateutil
